+++ src/main.lua
@@ -54,10 +54,25 @@ local function on_reload()
     if config.enabled == false then return end
 end
 
+local function on_ready_late()
+    if config.enabled == false then return end
+    import 'ready_late.lua'
+end
+
+local function on_reload_late()
+    if config.enabled == false then return end
+end
+
 -- this allows us to limit certain functions to not be reloaded.
-local loader = reload.auto_single()
+local loader = reload.auto_multiple()
 
 -- this runs only when modutil and the game's lua is ready
 modutil.once_loaded.game(function()
-    loader.load(on_ready, on_reload)
+    loader.load("early", on_ready, on_reload)
+end)
+
+mods.on_all_mods_loaded(function()
+	modutil.once_loaded.game(function()
+		loader.load("late", on_ready_late, on_reload_late)
+	end)
 end)